LoginHistoriesで直接hookを定義・ログイン履歴のhandlerを修正
