+++ tests/pages/api/user/login_history.test.ts
@@ -1,0 +1,5 @@
+import {testApiHandler} from 'next-test-api-route-handler';
+import login_historyHandler from '../../../../pages/api/user/login_history';
+import TestBase from '../../../../src/tests/base';
+
+describe('login_history', () => {});
